Refactored Clean Scripts
