+++ event_labelling/CodeStructure_Branching/clean_lable.py
@@ -13,159 +13,15 @@ Output: CLEAN_{team}_labels_branching_and_structure.csv with columns:
 """
 from __future__ import annotations
 import os
-import ast
-import pandas as pd
+import sys
 
+# Ensure project root is in path to allow imports from src/process_model
+project_root = os.path.abspath(os.path.join(os.path.dirname(__file__), '..', '..'))
+if project_root not in sys.path:
+    sys.path.insert(0, project_root)
 
-# PR-level label types to keep (filter out per-file labels)
-PR_LEVEL_LABELS = {
-    "Branch Name",
-    "Features Per Branch",
-    "Repository Status",
-    "Feature Size",
-    "Refactor Size",
-    "PR Status",
-    "Merge State"
-}
-
-
-def _parse_event_cell(ev) -> str:
-    """
-    Parse the event cell which might be a string or list.
-    Returns the event as a clean string.
-    """
-    if ev is None or (isinstance(ev, float) and pd.isna(ev)):
-        return ""
-
-    if isinstance(ev, str):
-        s = ev.strip()
-        # If it looks like a list literal, parse it
-        if s.startswith("[") and s.endswith("]"):
-            try:
-                parsed = ast.literal_eval(s)
-                if isinstance(parsed, list) and len(parsed) > 0:
-                    return str(parsed[0])
-            except Exception:
-                pass
-        return s
-
-    if isinstance(ev, list) and len(ev) > 0:
-        return str(ev[0])
-
-    return str(ev)
-
-
-def _pick_timestamp(row: pd.Series) -> str | None:
-    """
-    Timestamp selection logic:
-    - For Merge State: use merged_at
-    - For everything else: use created_at
-    """
-    main_label = row.get("main_label", "")
-    
-    # For Merge State, prefer merged_at
-    if main_label == "Merge State":
-        val = row.get("merged_at", None)
-        if val is not None and not (isinstance(val, float) and pd.isna(val)) and str(val).strip():
-            dt = pd.to_datetime(val, errors="coerce", utc=True)
-            if not pd.isna(dt):
-                return dt.strftime("%Y-%m-%dT%H:%M:%SZ")
-    
-    # Default to created_at
-    val = row.get("created_at", None)
-    if val is None or (isinstance(val, float) and pd.isna(val)) or not str(val).strip():
-        return None
-    
-    dt = pd.to_datetime(val, errors="coerce", utc=True)
-    if pd.isna(dt):
-        return str(val)
-    
-    return dt.strftime("%Y-%m-%dT%H:%M:%SZ")
-
-
-def create_clean_branching_label_csv(
-    input_csv_path: str, 
-    output_csv_path: str | None = None,
-    include_main_label: bool = False
-) -> str:
-    """
-    Reads {team}_labels_branching_and_structure.csv and creates a clean version
-    with only PR-level events (no per-file Feature/Refactor Size details).
-    
-    Args:
-        input_csv_path: Path to the full branching labels CSV
-        output_csv_path: Optional output path. If None, uses clean/ subfolder with CLEAN_ prefix
-        include_main_label: If True, includes main_label column in output
-    
-    Returns:
-        Path to the created clean CSV file
-    """
-    if not os.path.exists(input_csv_path):
-        raise FileNotFoundError(f"Input file not found: {input_csv_path}")
-
-    df = pd.read_csv(input_csv_path)
-    
-    # Validate required columns
-    required = {"pr_id", "event", "main_label", "created_at"}
-    missing = required - set(df.columns)
-    if missing:
-        raise KeyError(f"Missing required columns in {input_csv_path}: {sorted(missing)}")
-
-    # Default output path: clean/ subfolder with CLEAN_ prefix
-    if output_csv_path is None:
-        folder = os.path.dirname(input_csv_path)
-        clean_folder = os.path.join(folder, "clean")
-        os.makedirs(clean_folder, exist_ok=True)
-        base = os.path.basename(input_csv_path)
-        output_csv_path = os.path.join(clean_folder, f"CLEAN_{base}")
-
-    print(f"[INFO] Input: {input_csv_path}")
-    print(f"[INFO] Total rows: {len(df)}")
-    
-    # Filter to keep only PR-level labels
-    pr_level_df = df[df["main_label"].isin(PR_LEVEL_LABELS)].copy()
-    print(f"[INFO] PR-level rows after filtering: {len(pr_level_df)}")
-    
-    # Show breakdown by label type
-    print("[INFO] Breakdown by label type:")
-    for label_type in PR_LEVEL_LABELS:
-        count = (pr_level_df["main_label"] == label_type).sum()
-        if count > 0:
-            print(f"  - {label_type}: {count}")
-    
-    # Build output rows
-    out_rows = []
-    for _, row in pr_level_df.iterrows():
-        event = _parse_event_cell(row.get("event"))
-        ts = _pick_timestamp(row)
-        
-        out_row = {
-            "pr_id": row.get("pr_id"),
-            "timestamp": ts,
-            "event": event,
-        }
-        
-        if include_main_label:
-            out_row["main_label"] = row.get("main_label")
-        
-        out_rows.append(out_row)
-    
-    # Create output DataFrame
-    if include_main_label:
-        out_df = pd.DataFrame(out_rows, columns=["pr_id", "timestamp", "event", "main_label"])
-    else:
-        out_df = pd.DataFrame(out_rows, columns=["pr_id", "timestamp", "event"])
-    
-    # Ensure output directory exists
-    os.makedirs(os.path.dirname(output_csv_path), exist_ok=True)
-    
-    # Save to CSV
-    out_df.to_csv(output_csv_path, index=False)
-    
-    print(f"[SUCCESS] Clean labels saved to: {output_csv_path}")
-    print(f"[INFO] Output rows: {len(out_df)}")
-    
-    return output_csv_path
+# Import from the new utility
+from process_model.clean import create_clean_branching_label_csv
 
 
 def process_all_teams(base_folder: str = "data/graph_labels", include_main_label: bool = False):
@@ -197,7 +53,7 @@ def process_all_teams(base_folder: str = "data/graph_labels", include_main_label
     
     for file_path in files:
         team_name = os.path.basename(file_path).replace("_labels_branching_and_structure.csv", "")
-        print(f"\n[INFO] Processing: {team_name}")
+        print(f"\\n[INFO] Processing: {team_name}")
         print("-" * 70)
         
         try:
@@ -205,18 +61,17 @@ def process_all_teams(base_folder: str = "data/graph_labels", include_main_label
         except Exception as e:
             print(f"[ERROR] Failed to process {team_name}: {e}")
         
-    print("\n" + "=" * 70)
+    print("\\n" + "=" * 70)
     print("[COMPLETE] All files processed")
 
 
 if __name__ == "__main__":
-    import sys
     import glob
     
     # Usage examples:
-    # python get_clean_branching_label.py                                    # Process all *_labels_branching_and_structure.csv in current directory
-    # python get_clean_branching_label.py /path/to/file.csv                  # Process single file
-    # python get_clean_branching_label.py /path/to/folder --include-label    # Process folder with main_label column
+    # python clean_lable.py                                    # Process all *_labels_branching_and_structure.csv in current directory
+    # python clean_lable.py /path/to/file.csv                  # Process single file
+    # python clean_lable.py /path/to/folder --include-label    # Process folder with main_label column
     
     include_main_label = "--include-label" in sys.argv
     
@@ -259,14 +114,14 @@ if __name__ == "__main__":
             print("Search locations:")
             for path in search_paths:
                 print(f"  - {path}")
-            print("\nUsage: python get_clean_branching_label.py [file_or_folder]")
+            print("\\nUsage: python clean_lable.py [file_or_folder]")
         else:
             print(f"[INFO] Found {len(found_files)} file(s) to process")
             print("=" * 70)
             
             for file_path in found_files:
                 team_name = os.path.basename(file_path).replace("_labels_branching_and_structure.csv", "")
-                print(f"\n[INFO] Processing: {team_name}")
+                print(f"\\n[INFO] Processing: {team_name}")
                 print("-" * 70)
                 
                 try:
@@ -281,5 +136,5 @@ if __name__ == "__main__":
                     import traceback
                     traceback.print_exc()
             
-            print("\n" + "=" * 70)
+            print("\\n" + "=" * 70)
             print("[COMPLETE] All files processed")
